Remove old files, add new nuxt files
