+++ .gitignore
@@ -1,8 +1,9 @@
-*.log
-.cache
-.DS_Store
-src/.temp
 node_modules
-dist
+*.log*
+.nuxt
+.nitro
+.cache
+.output
 .env
-.env.*
+dist
+.DS_Store
