+++ .tool-versions
@@ -1,1 +1,1 @@
-nodejs 14.17.6
+nodejs 18.0.0
