+++ components/Animation.vue
@@ -1,0 +1,495 @@
+<template>
+    <div id="gears" ref="gearsContainer">
+        <svg id="svg2" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:cc="http://creativecommons.org/ns#" xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:svg="http://www.w3.org/2000/svg" xmlns="http://www.w3.org/2000/svg" version="1.1" viewbox="0 0 744.09448819 1052.3622047">
+            <path id="path6863" style="fill:none;stroke:#000000;stroke-width:1" d="m 215.42169,-22.08672 12.31225,-1.26235 14.95886,2.1162 27.83871,12.00872 0,18.4483 -27.83871,12.00872 -14.95886,2.1162 -12.31225,-1.26235 -5.43742,30.83711 12.00149,3.02481 13.33294,7.10482 22.05261,20.8059 -6.30969,17.33573 -30.26705,1.76311 -14.78052,-3.12765 -11.13799,-5.39726 -15.6564,27.11769 10.24315,6.94715 10.09889,11.23648 13.60663,27.09359 -11.85834,14.13221 -29.04474,-8.69516 -12.81943,-7.99427 -8.62031,-8.88118 -23.987,20.12749 7.24935,10.03155 5.64675,14.01286 3.51949,30.11339 -15.97669,9.22415 -24.31921,-18.10467 -9.31212,-11.89666 -5.0629,-11.2939 -29.42442,10.70962 3.38117,11.906 0.51353,15.09907 -6.99214,29.50107 -18.16803,3.20352 -16.66042,-25.33049 -4.68163,-14.36413 -0.89483,-12.3444 -31.31282,0 -0.89483,12.3444 -4.68163,14.36413 -16.66042,25.33049 -18.16803,-3.20352 -6.99214,-29.50107 0.51353,-15.09907 3.38117,-11.906 -29.42442,-10.70962 -5.0629,11.2939 -9.31212,11.89666 -24.31921,18.10467 -15.97669,-9.22415 3.51949,-30.11339 5.64675,-14.01286 7.24935,-10.03155 -23.987,-20.12749 -8.62031,8.88118 -12.81943,7.99427 -29.04474,8.69516 -11.85834,-14.13221 13.60663,-27.09359 10.09889,-11.23648 10.24315,-6.94715 -15.6564,-27.11769 -11.13799,5.39726 -14.78052,3.12765 -30.26705,-1.76311 -6.30969,-17.33573 22.05261,-20.8059 13.33294,-7.10482 12.00149,-3.02481 -5.43742,-30.83711 -12.31225,1.26235 -14.95886,-2.1162 -27.83871,-12.00872 0,-18.4483 27.83871,-12.00872 14.95886,-2.1162 12.31225,1.26235 5.43742,-30.83711 -12.00149,-3.02481 -13.33294,-7.10482 -22.05261,-20.8059 6.30969,-17.33573 30.26705,-1.76311 14.78052,3.12765 11.13799,5.39726 15.6564,-27.11769 -10.24315,-6.94715 -10.09889,-11.23648 -13.60663,-27.09359 11.85834,-14.13221 29.04474,8.69516 12.81943,7.99427 8.62031,8.88118 23.987,-20.12749 -7.24935,-10.03155 -5.64675,-14.01286 -3.51949,-30.11339 15.97669,-9.22415 24.31921,18.10467 9.31212,11.89666 5.0629,11.2939 29.42442,-10.70962 -3.38117,-11.906 -0.51353,-15.09907 6.99214,-29.50107 18.16803,-3.20352 16.66042,25.33049 4.68163,14.36413 0.89483,12.3444 31.31282,0 0.89483,-12.3444 4.68163,-14.36413 16.66042,-25.33049 18.16803,3.20352 6.99214,29.50107 -0.51353,15.09907 -3.38117,11.906 29.42442,10.70962 5.0629,-11.2939 9.31212,-11.89666 24.31921,-18.10467 15.97669,9.22415 -3.51949,30.11339 -5.64675,14.01286 -7.24935,10.03155 23.987,20.12749 8.62031,-8.88118 12.81943,-7.99427 29.04474,-8.69516 11.85834,14.13221 -13.60663,27.09359 -10.09889,11.23648 -10.24315,6.94715 15.6564,27.11769 11.13799,-5.39726 14.78052,-3.12765 30.26705,1.76311 6.30969,17.33573 -22.05261,20.8059 -13.33294,7.10482 -12.00149,3.02481 z"></path>
+        </svg>
+        <svg id="svg4338" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:cc="http://creativecommons.org/ns#" xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:svg="http://www.w3.org/2000/svg" xmlns="http://www.w3.org/2000/svg" width="210mm" height="297mm" viewbox="0 0 744.09448819 1052.3622047">
+            <path id="path6936" style="fill:none;stroke:#000000;stroke-width:1" d="m 459.39171,-26.76653 27.38425,5.51343 27.43171,11.06592 0,20.37436 -27.43171,11.06592 -27.38425,5.51343 -2.33123,26.64612 26.01082,10.18489 25.09339,15.66128 -3.53797,20.06481 -28.93655,6.13434 -27.92561,0.67445 -6.92287,25.83649 23.84707,14.54689 21.99261,19.78077 -6.96844,19.14563 -29.56214,1.01636 -27.61848,-4.18503 -11.30415,24.24183 20.95873,18.4669 18.2236,23.29922 -10.18717,17.64471 -29.28953,-4.13249 -26.47216,-8.91735 -15.34197,21.9106 17.43358,21.82579 13.90088,26.10975 -13.09638,15.60766 -28.12695,-9.15578 -24.52152,-13.37872 -18.91362,18.91362 13.37872,24.52152 9.15578,28.12695 -15.60766,13.09638 -26.10975,-13.90088 -21.82579,-17.43358 -21.9106,15.34197 8.91735,26.47216 4.13249,29.28953 -17.64471,10.18717 -23.29922,-18.2236 -18.4669,-20.95873 -24.24183,11.30415 4.18503,27.61848 -1.01636,29.56214 -19.14563,6.96844 -19.78077,-21.99261 -14.54689,-23.84707 -25.83649,6.92287 -0.67445,27.92561 -6.13434,28.93655 -20.06481,3.53797 -15.66128,-25.09339 -10.18489,-26.01082 -26.64612,2.33123 -5.51343,27.38425 -11.06592,27.43171 -20.37436,0 -11.06592,-27.43171 -5.51343,-27.38425 -26.64612,-2.33123 -10.18489,26.01082 -15.66128,25.09339 -20.06481,-3.53797 -6.13434,-28.93655 -0.67445,-27.92561 -25.83649,-6.92287 -14.54689,23.84707 -19.78077,21.99261 -19.14563,-6.96844 -1.01636,-29.56214 4.18503,-27.61848 -24.24183,-11.30415 -18.4669,20.95873 -23.29922,18.2236 -17.64471,-10.18717 4.13249,-29.28953 8.91735,-26.47216 -21.9106,-15.34197 -21.82579,17.43358 -26.10975,13.90088 -15.60766,-13.09638 9.15578,-28.12695 13.37872,-24.52152 -18.91362,-18.91362 -24.52152,13.37872 -28.12695,9.15578 -13.09638,-15.60766 13.90088,-26.10975 17.43358,-21.82579 -15.34197,-21.9106 -26.47216,8.91735 -29.28953,4.13249 -10.18717,-17.64471 18.2236,-23.29922 20.95873,-18.4669 -11.30415,-24.24183 -27.61848,4.18503 -29.56214,-1.01636 -6.96844,-19.14563 21.99261,-19.78077 23.84707,-14.54689 -6.92287,-25.83649 -27.92561,-0.67445 -28.93655,-6.13434 -3.53797,-20.06481 25.09339,-15.66128 26.01082,-10.18489 -2.33123,-26.64612 -27.38425,-5.51343 -27.43171,-11.06592 0,-20.37436 27.43171,-11.06592 27.38425,-5.51343 2.33123,-26.64612 -26.01082,-10.18489 -25.09339,-15.66128 3.53797,-20.06481 28.93655,-6.13434 27.92561,-0.67445 6.92287,-25.83649 -23.84707,-14.54689 -21.99261,-19.78077 6.96844,-19.14563 29.56214,-1.01636 27.61848,4.18503 11.30415,-24.24183 -20.95873,-18.4669 -18.2236,-23.29922 10.18717,-17.64471 29.28953,4.13249 26.47216,8.91735 15.34197,-21.9106 -17.43358,-21.82579 -13.90088,-26.10975 13.09638,-15.60766 28.12695,9.15578 24.52152,13.37872 18.91362,-18.91362 -13.37872,-24.52152 -9.15578,-28.12695 15.60766,-13.09638 26.10975,13.90088 21.82579,17.43358 21.9106,-15.34197 -8.91735,-26.47216 -4.13249,-29.28953 17.64471,-10.18717 23.29922,18.2236 18.4669,20.95873 24.24183,-11.30415 -4.18503,-27.61848 1.01636,-29.56214 19.14563,-6.96844 19.78077,21.99261 14.54689,23.84707 25.83649,-6.92287 0.67445,-27.92561 6.13434,-28.93655 20.06481,-3.53797 15.66128,25.09339 10.18489,26.01082 26.64612,-2.33123 5.51343,-27.38425 11.06592,-27.43171 20.37436,0 11.06592,27.43171 5.51343,27.38425 26.64612,2.33123 10.18489,-26.01082 15.66128,-25.09339 20.06481,3.53797 6.13434,28.93655 0.67445,27.92561 25.83649,6.92287 14.54689,-23.84707 19.78077,-21.99261 19.14563,6.96844 1.01636,29.56214 -4.18503,27.61848 24.24183,11.30415 18.4669,-20.95873 23.29922,-18.2236 17.64471,10.18717 -4.13249,29.28953 -8.91735,26.47216 21.9106,15.34197 21.82579,-17.43358 26.10975,-13.90088 15.60766,13.09638 -9.15578,28.12695 -13.37872,24.52152 18.91362,18.91362 24.52152,-13.37872 28.12695,-9.15578 13.09638,15.60766 -13.90088,26.10975 -17.43358,21.82579 15.34197,21.9106 26.47216,-8.91735 29.28953,-4.13249 10.18717,17.64471 -18.2236,23.29922 -20.95873,18.4669 11.30415,24.24183 27.61848,-4.18503 29.56214,1.01636 6.96844,19.14563 -21.99261,19.78077 -23.84707,14.54689 6.92287,25.83649 27.92561,0.67445 28.93655,6.13434 3.53797,20.06481 -25.09339,15.66128 -26.01082,10.18489 z"></path>
+        </svg>
+        <svg id="svg5002" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:cc="http://creativecommons.org/ns#" xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:svg="http://www.w3.org/2000/svg" xmlns="http://www.w3.org/2000/svg" version="1.1" viewbox="0 0 744.09448819 1052.3622047">
+            <path id="path6992" style="fill:none;stroke:#000000;stroke-width:1" d="m 676.13057,-28.28177 27.33903,7.02516 27.31414,10.7304 0,21.05242 -27.31414,10.7304 -27.33903,7.02516 -1.52077,25.1412 26.29291,10.2693 25.82159,13.94452 -2.53759,20.89892 -28.4084,7.3598 -27.98649,3.67858 -4.54011,24.7746 24.86338,13.36367 23.95249,16.9553 -5.03817,20.44067 -29.0884,3.88189 -28.22584,0.27837 -7.49326,24.0467 23.07129,16.26319 21.73412,19.71883 -7.46529,19.68435 -29.34423,0.34737 -28.05359,-3.12592 -10.33713,22.96817 20.94275,18.92555 19.19881,22.19481 -9.78354,18.64099 -29.17215,-3.19221 -27.47226,-6.48462 -13.03027,21.5547 18.50884,21.31193 16.38354,24.34715 -11.95914,17.3258 -28.57466,-6.68525 -26.49033,-9.74875 -15.53339,19.82691 15.80501,23.38754 13.32936,26.14445 -13.96033,15.75796 -27.5605,-10.08081 -25.1221,-12.87072 -17.81001,17.81001 12.87072,25.1221 10.08081,27.5605 -15.75796,13.96033 -26.14445,-13.32936 -23.38754,-15.80501 -19.82691,15.53339 9.74875,26.49033 6.68525,28.57466 -17.3258,11.95914 -24.34715,-16.38354 -21.31193,-18.50884 -21.5547,13.03027 6.48462,27.47226 3.19221,29.17215 -18.64099,9.78354 -22.19481,-19.19881 -18.92555,-20.94275 -22.96817,10.33713 3.12592,28.05359 -0.34737,29.34423 -19.68435,7.46529 -19.71883,-21.73412 -16.26319,-23.07129 -24.0467,7.49326 -0.27837,28.22584 -3.88189,29.0884 -20.44067,5.03817 -16.9553,-23.95249 -13.36367,-24.86338 -24.7746,4.54011 -3.67858,27.98649 -7.3598,28.4084 -20.89892,2.53759 -13.94452,-25.82159 -10.2693,-26.29291 -25.1412,1.52077 -7.02516,27.33903 -10.7304,27.31414 -21.05242,0 -10.7304,-27.31414 -7.02516,-27.33903 -25.1412,-1.52077 -10.2693,26.29291 -13.94452,25.82159 -20.89892,-2.53759 -7.3598,-28.4084 -3.67858,-27.98649 -24.7746,-4.54011 -13.36367,24.86338 -16.9553,23.95249 -20.44067,-5.03817 -3.88189,-29.0884 -0.27837,-28.22584 -24.0467,-7.49326 -16.26319,23.07129 -19.71883,21.73412 -19.68435,-7.46529 -0.34737,-29.34423 3.12592,-28.05359 -22.96817,-10.33713 -18.92555,20.94275 -22.19481,19.19881 -18.64099,-9.78354 3.19221,-29.17215 6.48462,-27.47226 -21.5547,-13.03027 -21.31193,18.50884 -24.34715,16.38354 -17.3258,-11.95914 6.68525,-28.57466 9.74875,-26.49033 -19.82691,-15.53339 -23.38754,15.80501 -26.14445,13.32936 -15.75796,-13.96033 10.08081,-27.5605 12.87072,-25.1221 -17.81001,-17.81001 -25.1221,12.87072 -27.5605,10.08081 -13.96033,-15.75796 13.32936,-26.14445 15.80501,-23.38754 -15.53339,-19.82691 -26.49033,9.74875 -28.57466,6.68525 -11.95914,-17.3258 16.38354,-24.34715 18.50884,-21.31193 -13.03027,-21.5547 -27.47226,6.48462 -29.17215,3.19221 -9.78354,-18.64099 19.19881,-22.19481 20.94275,-18.92555 -10.33713,-22.96817 -28.05359,3.12592 -29.34423,-0.34737 -7.46529,-19.68435 21.73412,-19.71883 23.07129,-16.26319 -7.49326,-24.0467 -28.22584,-0.27837 -29.0884,-3.88189 -5.03817,-20.44067 23.95249,-16.9553 24.86338,-13.36367 -4.54011,-24.7746 -27.98649,-3.67858 -28.4084,-7.3598 -2.53759,-20.89892 25.82159,-13.94452 26.29291,-10.2693 -1.52077,-25.1412 -27.33903,-7.02516 -27.31414,-10.7304 0,-21.05242 27.31414,-10.7304 27.33903,-7.02516 1.52077,-25.1412 -26.29291,-10.2693 -25.82159,-13.94452 2.53759,-20.89892 28.4084,-7.3598 27.98649,-3.67858 4.54011,-24.7746 -24.86338,-13.36367 -23.95249,-16.9553 5.03817,-20.44067 29.0884,-3.88189 28.22584,-0.27837 7.49326,-24.0467 -23.07129,-16.26319 -21.73412,-19.71883 7.46529,-19.68435 29.34423,-0.34737 28.05359,3.12592 10.33713,-22.96817 -20.94275,-18.92555 -19.19881,-22.19481 9.78354,-18.64099 29.17215,3.19221 27.47226,6.48462 13.03027,-21.5547 -18.50884,-21.31193 -16.38354,-24.34715 11.95914,-17.3258 28.57466,6.68525 26.49033,9.74875 15.53339,-19.82691 -15.80501,-23.38754 -13.32936,-26.14445 13.96033,-15.75796 27.5605,10.08081 25.1221,12.87072 17.81001,-17.81001 -12.87072,-25.1221 -10.08081,-27.5605 15.75796,-13.96033 26.14445,13.32936 23.38754,15.80501 19.82691,-15.53339 -9.74875,-26.49033 -6.68525,-28.57466 17.3258,-11.95914 24.34715,16.38354 21.31193,18.50884 21.5547,-13.03027 -6.48462,-27.47226 -3.19221,-29.17215 18.64099,-9.78354 22.19481,19.19881 18.92555,20.94275 22.96817,-10.33713 -3.12592,-28.05359 0.34737,-29.34423 19.68435,-7.46529 19.71883,21.73412 16.26319,23.07129 24.0467,-7.49326 0.27837,-28.22584 3.88189,-29.0884 20.44067,-5.03817 16.9553,23.95249 13.36367,24.86338 24.7746,-4.54011 3.67858,-27.98649 7.3598,-28.4084 20.89892,-2.53759 13.94452,25.82159 10.2693,26.29291 25.1412,-1.52077 7.02516,-27.33903 10.7304,-27.31414 21.05242,0 10.7304,27.31414 7.02516,27.33903 25.1412,1.52077 10.2693,-26.29291 13.94452,-25.82159 20.89892,2.53759 7.3598,28.4084 3.67858,27.98649 24.7746,4.54011 13.36367,-24.86338 16.9553,-23.95249 20.44067,5.03817 3.88189,29.0884 0.27837,28.22584 24.0467,7.49326 16.26319,-23.07129 19.71883,-21.73412 19.68435,7.46529 0.34737,29.34423 -3.12592,28.05359 22.96817,10.33713 18.92555,-20.94275 22.19481,-19.19881 18.64099,9.78354 -3.19221,29.17215 -6.48462,27.47226 21.5547,13.03027 21.31193,-18.50884 24.34715,-16.38354 17.3258,11.95914 -6.68525,28.57466 -9.74875,26.49033 19.82691,15.53339 23.38754,-15.80501 26.14445,-13.32936 15.75796,13.96033 -10.08081,27.5605 -12.87072,25.1221 17.81001,17.81001 25.1221,-12.87072 27.5605,-10.08081 13.96033,15.75796 -13.32936,26.14445 -15.80501,23.38754 15.53339,19.82691 26.49033,-9.74875 28.57466,-6.68525 11.95914,17.3258 -16.38354,24.34715 -18.50884,21.31193 13.03027,21.5547 27.47226,-6.48462 29.17215,-3.19221 9.78354,18.64099 -19.19881,22.19481 -20.94275,18.92555 10.33713,22.96817 28.05359,-3.12592 29.34423,0.34737 7.46529,19.68435 -21.73412,19.71883 -23.07129,16.26319 7.49326,24.0467 28.22584,0.27837 29.0884,3.88189 5.03817,20.44067 -23.95249,16.9553 -24.86338,13.36367 4.54011,24.7746 27.98649,3.67858 28.4084,7.3598 2.53759,20.89892 -25.82159,13.94452 -26.29291,10.2693 z"></path>
+        </svg>
+    </div>
+    <div id="canvas-container" ref="canvasContainer"></div>
+</template>
+<script setup>
+import { ref, watchEffect } from 'vue'
+import Two from 'two.js'
+
+var two = new Two({
+    fullscreen: false,
+    type: Two.Types.svg
+})
+
+var theme = useState('theme');
+
+const canvasContainer = ref(null);
+const gearsContainer = ref(null);
+
+onMounted(() => {
+
+    var footer = document.getElementById('footer');
+    two.width = window.visualViewport.width;
+    two.height = window.visualViewport.height - footer.clientHeight - 2;
+
+    /*
+    * Main Groups
+    */
+
+    var background = two.makeGroup();
+    background.id = 'background';
+    var foreground = two.makeGroup();
+    foreground.id = "foreground";
+
+
+    /*
+    * Background Groups
+    */
+
+    var gadgetsGroup = two.makeGroup();
+    gadgetsGroup.addTo(background);
+    gadgetsGroup.id = 'gadget-group';
+
+    var thingsGroup = two.makeGroup();
+    thingsGroup.addTo(background);
+    thingsGroup.id = 'things-group';
+
+    var poetryGroup = two.makeGroup();
+    poetryGroup.addTo(background);
+    poetryGroup.id = 'poetry-group';
+
+    /*
+    * Foreground Groups
+    */
+
+    var textGroup = two.makeGroup();
+    textGroup.addTo(foreground);
+    textGroup.id = 'text-group';
+
+
+    /*
+    * Bounding Box Template
+    * Mostly to help transform and scale SVGs
+    */
+
+    for (var i = 0; i < background.children.length - 1; i++) {
+        var bb = two.makeRectangle(250, two.height / 2, 500, two.height);
+        bb.noStroke().noFill().visible = false;
+        background.children[i].add(bb);
+    }
+
+    /*
+    * Timer Line
+    */
+
+    var timerLine = two.makeLine(0, 187, 130, 187);
+    timerLine.visible = false;
+    timerLine.linewidth = 10;
+    timerLine.opacity = .4
+    textGroup.add(timerLine);
+
+
+    /*
+    * Text
+    */
+
+    var textStyles = {
+        family: "'Fjalla One', sans-serif",
+        weight: 400,
+        alignment: 'left',
+        visible: false
+    };
+
+    textStyles['size'] = 50;
+    textStyles['leading'] = 50;
+    var andy = two.makeText('Andy', 0, 20, textStyles);
+    var makes = two.makeText('Makes', 0, 90, textStyles);
+    var things = two.makeText('Things', 0, 160, textStyles);
+
+    textGroup.add(andy, makes, things);
+    textGroup.translation.set(16, 88);
+
+    var utils = {
+        randomNumber: function(max, min, whole = true) {
+            if (whole) {
+                return Math.floor(Math.random() * (max - min)) + min;
+            } else {
+                return Math.random() * (max - min) + min;
+            }
+        },
+        resize: function(background, foreground) {
+            background.scale = canvasContainer.value.clientHeight / 540;
+            console.log(background.scale)
+        },
+        fade: function(objects, direction, frameCount, start, end) {
+            if (frameCount >= start && frameCount <= end){
+                for (var i = 0; i < objects.length; i++) {
+                    if (direction == 1) {
+                        if (objects[i].opacity < 1) {
+                            objects[i].opacity += .1;
+                        } else if (objects[i].opacity > 1){
+                            objects[i].opacity = Math.floor(objects[i].opacity)
+                        }
+                    } else {
+                        if (objects[i].opacity > 0) {
+                            objects[i].opacity -= .1;
+                        } else if (objects[i].opacity < 0){
+                            objects[i].opacity = Math.floor(objects[i].opacity)
+                        }
+                    }
+                }
+            }
+        },
+        handleTimer: function(timeOut, frameCount, timerLine) {
+            //var timerWidth = things.getBoundingClientRect().width;
+            var modulo = frameCount % timeOut;
+            var pctTime = modulo / timeOut;
+            var pctBar = pctTime * 130;
+            timerLine.vertices[1].x = 130 - pctBar;
+        },
+        setThemeColor: function(theme) {
+            if (theme) {
+                andy.fill = 'white';
+                makes.fill = 'white';
+                things.fill = 'white';
+                timerLine.noFill().stroke = 'white';
+            } else {
+                andy.fill = '#212121';
+                makes.fill = '#212121';
+                things.fill = '#212121';
+                timerLine.noFill().stroke = '#212121';
+            }
+        }
+    }
+
+    /*
+    * Beams
+    */
+
+    var beams = {
+        group: thingsGroup,
+        paths: [],
+        animate: function(frameCount, startFrame, endFrame) {
+            if (frameCount >= startFrame && frameCount <= endFrame && frameCount % 3 == 0){
+                this.populate(10);
+
+                // Animate beams
+                for (var i = 0; i < this.paths.length; i++) {
+                    this.paths[i].controller.frameCount++;
+                    if (this.paths[i].controller.frameCount > this.paths[i].controller.length / 2) {
+                    this.paths[i].opacity -= this.paths[i].controller.speed;
+                    } else if (this.paths[i].opacity <= this.paths[i].controller.peak) {
+                    this.paths[i].opacity += this.paths[i].controller.speed;
+                    }
+                    this.paths[i].vertices[1].x += this.paths[i].controller.travel;
+                    this.paths[i].vertices[2].x += this.paths[i].controller.travel;
+
+                    if (this.paths[i].opacity <= 0) {
+                    this.paths[i].opacity = 0;
+                    this.paths[i].controller = this.createController();
+                    }
+                }
+            }
+        },
+        create: function() {
+            var beam = two.makePath(300, 0, utils.randomNumber(400, 0), two.height, utils.randomNumber(400, 0), two.height)
+            beam.noStroke().fill = '#FFEB3B';
+            beam.opacity = 0;
+            beam.controller = this.createController();
+            this.paths.push(beam);
+            this.group.add(beam)
+        },
+        createController: function() {
+            return {
+            peak: utils.randomNumber(.7, .1, false),
+            length: utils.randomNumber(180, 30, false),
+            speed: utils.randomNumber(.1, .01, false),
+            travel: utils.randomNumber(.5, -.5, false),
+            frameCount: 0
+            }
+        },
+        populate: function(target) {
+            if (this.paths.length < target) {
+            var current = this.paths.length;
+            for (var i = current; i < target; i++) {
+                this.create();
+            }
+            }
+        },
+        destroy: function() {
+            this.group.remove(this.paths);
+            this.paths = [];
+        }
+    }
+
+
+    /*
+    * Poetry
+    */
+
+    var poetry = {
+        group: poetryGroup,
+        paths: [],
+        line: {
+            x: 20,
+            y: -30,
+            height: 40,
+            minWidth: 100,
+            maxWidth: 380
+        },
+        step: 60,
+
+        create: function(i) {
+            var width = utils.randomNumber(this.line.maxWidth, this.line.minWidth);
+            var line = two.makeRectangle(
+                (width / 2) + this.line.x,
+                this.line.y + (i * this.step),
+                width,
+                this.line.height
+            );
+            line.rotation = utils.randomNumber(5, -5) * (Math.PI / 180);
+            line.noStroke().fill = 'red';
+            line.opacity = .75;
+            for (var v = 0; v < 4; v++) {
+                line.vertices[v].ogx = line.vertices[v].x;
+            }
+            this.paths.push(line);
+            this.group.add(line);
+        },
+        populate: function() {
+            var target = Math.floor(document.documentElement.clientHeight / this.step);
+            var current = this.paths.length;
+            if (current < target) {
+                for (var i = current; i <= target; i++) {
+                    this.create(i);
+                }
+            }
+        },
+        animate: function(frameCount, startFrame, endFrame) {
+            if (frameCount >= startFrame && frameCount <= endFrame && frameCount % 3 == 0) {
+                this.populate();
+
+                for (var i = 0; i < this.paths.length; i++) {
+                    for (var v = 0; v < 4; v++) {
+                        if (this.paths[i].vertices[v].hasOwnProperty('jitter')) {
+                            this.paths[i].vertices[v].x += this.paths[i].vertices[v].dir;
+                            if (this.paths[i].vertices[v].x == this.paths[i].vertices[v].ogx + this.paths[i].vertices[v].jitter) {
+                                this.paths[i].vertices[v].dir = this.paths[i].vertices[v].dir * -1;
+                            }
+                            if (this.paths[i].vertices[v].x == this.paths[i].vertices[v].ogx) {
+                                delete this.paths[i].vertices[v].jitter;
+                                delete this.paths[i].vertices[v].dir;
+                            }
+                        } else {
+                            var jitter = 0;
+                            while (jitter == 0) {
+                                jitter = utils.randomNumber(6, -6);
+                            }
+                            this.paths[i].vertices[v].jitter = jitter;
+                            this.paths[i].vertices[v].dir = this.paths[i].vertices[v].jitter > 0 ? 1 : -1;
+                        }
+                    }
+                }
+            }
+        },
+        destroy: function() {
+            this.group.remove(this.paths);
+            this.paths = [];
+        }
+    }
+
+
+    /*
+    * Gears
+    */
+
+    var gears = {
+        group: gadgetsGroup,
+        diametralPitch: Math.PI / 24,
+        colors: ['#D500F9', '#00B0FF', '#FF1744'],
+        teeth: [18, 36, 52],
+        assemblies: [],
+        gears: [],
+        scaleRatio: undefined,
+        pinionStep: .25,
+        animate: function(frameCount, startFrame, endFrame) {
+            if (frameCount >= startFrame && frameCount <= endFrame) {
+                if (this.assemblies.length != gearsContainer.value.children.length) {
+                    this.populate();
+                    this.place();
+                }
+
+                // RPM = Step * FPS * Seconds
+                var pinionRPM = this.pinionStep * 60 * 60;
+
+                this.gears[0].rotation += this.pinionStep * (Math.PI / 180);
+                if (this.gears[0].rotation >= 360) {
+                    this.gears[0].rotation = 0;
+                }
+
+                // s1 * teeth1 = s2 * teeth2
+                this.gear1Step = ((pinionRPM * this.gears[0].teeth) / this.gears[1].teeth) / 60 / 60;
+                this.gears[1].rotation -= this.gear1Step * (Math.PI / 180);
+                if (this.gears[1].rotation >= 360) {
+                    this.gears[1].rotation = 0;
+                }
+
+                this.gear2Step = ((pinionRPM * this.gears[0].teeth) / this.gears[2].teeth) / 60 / 60;
+                this.gears[2].rotation -= this.gear2Step * (Math.PI / 180);
+                if (this.gears[2].rotation >= 360) {
+                    this.gears[2].rotation = 0;
+                }
+            }
+        },
+        create: function(i, svg) {
+            var assembly = two.makeGroup();
+            this.assemblies.push(assembly);
+            var gear = two.interpret(svg).center();
+            assembly.add(gear);
+            this.gears.push(gear);
+            if (this.scaleRatio == undefined) {
+                this.scaleRatio = 200 / gear.getBoundingClientRect().width;
+            }
+            gear.scale = this.scaleRatio;
+            gear.teeth = this.teeth[i];
+            gear.diametralPitch = (gear.teeth + 2) / gear.getBoundingClientRect().width;
+            gear.pitchDiameter = gear.getBoundingClientRect().width - (2 / gear.diametralPitch);
+            gear.fill = 'transparent';
+            gear.stroke = this.colors[i];
+            gear.linewidth = 5;
+            gear.opacity = .7;
+
+            var circle = two.makeCircle(0, 0, (gear.getBoundingClientRect().width - 60) / 2);
+            assembly.add(circle);
+            circle.fill =  'transparent'; // colors[i];
+            circle.stroke = this.colors[i];
+            circle.linewidth = 15;
+            this.group.add(assembly);
+        },
+        populate: function() {
+            var svgs = gearsContainer.value.children;
+            for (var i = 0; i < svgs.length; i++) {
+                this.create(i, svgs[i]);
+            }
+        },
+        place: function() {
+            this.assemblies[1].translation.set(50, 547.5);
+            this.assemblies[0].translation.set(
+                this.assemblies[1].translation.x + ((this.assemblies[1].children[0].pitchDiameter + this.assemblies[0].children[0].pitchDiameter) / 2) * Math.cos(-(5 * (360 / 36)) * (Math.PI / 180)),
+                this.assemblies[1].translation.y + ((this.assemblies[1].children[0].pitchDiameter + this.assemblies[0].children[0].pitchDiameter) / 2) * Math.sin(-(5 * (360 / 36)) * (Math.PI / 180))
+            );
+
+        this.assemblies[2].translation.set(
+                this.assemblies[0].translation.x + ((this.assemblies[2].children[0].pitchDiameter + this.assemblies[0].children[0].pitchDiameter) / 2) * Math.cos((13 * (360 / 18) - .25) * (Math.PI / 180)),
+                this.assemblies[0].translation.y + ((this.assemblies[2].children[0].pitchDiameter + this.assemblies[0].children[0].pitchDiameter) / 2) * Math.sin((13 * (360 / 18) - .25) * (Math.PI / 180))
+            );
+        },
+        destroy: function() {
+            this.group.remove(this.assemblies);
+            this.assemblies = [];
+            this.gears = [];
+        }
+    }
+
+    for (var i = 0; i < background.children.length; i++) {
+        background.children[i].visible = false;
+        background.children[i].opacity = 0;
+    }
+
+    //utils.resize(gadgetsGroup);
+    gadgetsGroup.scale = background.scale;
+    var iframeCount = 0;
+    window.addEventListener('resize', utils.resize(background))
+    two.bind('resize', function() {
+        utils.resize(background);
+    })
+    .bind('update', function(frameCount) {
+        if (frameCount == 20) {
+            andy.visible = true;
+        }
+        if (frameCount == 50) {
+            makes.visible = true;
+        }
+        if (frameCount == 80) {
+            timerLine.visible = true;
+            utils.fade([timerLine], 1, frameCount, 80, 89);
+        }
+
+        if (frameCount >= 80) {
+            timerLine.visible = true;
+            utils.handleTimer(450, iframeCount, timerLine);
+            beams.animate(iframeCount, 0, 449);
+            poetry.animate(iframeCount, 450, 899);
+            gears.animate(iframeCount, 900, 1349);
+
+            if (iframeCount == 0) {
+                things.visible = true;
+                things.value = 'Things';
+                thingsGroup.visible = true;
+            }
+            if (iframeCount == 449) {
+                thingsGroup.visible = false;
+                beams.destroy();
+            }
+            utils.fade([makes, things, thingsGroup], 1, iframeCount, 0, 9);
+            utils.fade([makes, things, thingsGroup], -1, iframeCount, 440, 449)
+
+            if (iframeCount == 450) {
+                poetryGroup.visible = true;
+                makes.value = 'Crafts'
+                things.value = 'Poems';
+            }
+            if (iframeCount == 899) {
+                poetryGroup.visible = false;
+                poetry.destroy();
+            }
+            utils.fade([makes, things, poetryGroup], 1, iframeCount, 450, 459);
+            utils.fade([makes, things, poetryGroup], -1, iframeCount, 890, 899);
+
+            if (iframeCount == 900) {
+                makes.value = 'Constructs';
+                things.value = 'Gadgets';
+                gadgetsGroup.visible = true;
+            }
+            if (iframeCount == 1349) {
+                gears.destroy();
+            }
+            utils.fade([makes, things, gadgetsGroup], 1, iframeCount, 900, 909);
+            utils.fade([makes, things, gadgetsGroup], -1, iframeCount, 1340, 1349);
+
+            iframeCount += 1;
+            if (iframeCount == 1350) {
+                iframeCount = 0;
+                makes.value = 'Makes';
+                things.value = 'Things';
+            }
+        }
+    }).play();
+
+    utils.setThemeColor(theme.value)
+
+    watchEffect(() => {
+        utils.setThemeColor(theme.value)
+    })
+
+    two.appendTo(canvasContainer.value);
+    two.trigger('resize');
+})
+
+onDeactivated(() => {
+    two.unbind("update");
+})
+</script>
+
+<style>
+#gears {
+    display: none;
+}
+
+#canvas-container {
+    min-width: 100%;
+    height: 100%;
+    min-height: 100%;
+    position: relative;
+}
+</style>
