+++ content/about.md
@@ -1,0 +1,25 @@
+---
+title: 'About'
+description: 'A brief bio and colophon'
+slug: about
+---
+
+## The Basics
+
+Andy Stevens is a technical writer by day, and a poet, web developer, and tinkerer by night. Andy holds a B.A. in Creative Writing from Eckerd College in St. Petersburg, FL, and an M.F.A in Creative Writing from University of Tampa.
+
+His poetry and writing have been published in a few places, but the internet being what it is, not all of it still exists. He is currently working on his first book-length collection. If you're a friend or family member, or an editor, and would like an advanced copy of the collection-in-progress, please reach out.
+
+## Social Media
+
+Andy can be reached on the following social media platforms:
+
+- Twitter: [@andystevensname](https://twitter.com/andystevensname)
+
+## Colophon
+
+This website is powered by [Nuxt.js 3](https://v3.nuxtjs.org/) acting as a static site generator, which employs the [Vue.js framework](https://vuejs.org/).
+
+The animation on the front page is powered by [two.js](https://two.js.org/), an SVG animation library.
+
+This site's code is available on [GitHub](https://github.com/andystevensname/andystevens.name).
