+++ content/writing.md
@@ -1,0 +1,27 @@
+---
+title: 'Writing'
+description: 'A list of publications, links, and other notable accomplishments.'
+slug: 'writing'
+---
+
+## Poetry
+
+### Publications
+
+- Eckerd Review - 2011
+- [Digital Americana](http://www.magcloud.com/browse/issue/387253) - Summer 2012 (published under Leroy van Stets)
+- [Barely South Review](http://issuu.com/barelysouth/docs/barely_south_review_fall_2013) - Fall 2013
+- [Prick of the Spindle](http://prickofthespindle.org/product/print-edition-issue-8/) - Issue 8, Spring/Summer 2015
+- [Bridge Eight](https://www.bridgeeight.com/poems-jack-freeman-andy-stevens/)
+
+### Awards and Other Recognition
+
+- 2009 ASPEC Mason Daly Poetry Award from Eckerd College
+- June 2013 Winner Tampa Bay Hotel / Plant Hall Beautiful-Weird-Cool-Spooky Old Place Writing Contest at University of Tampa
+- Finalist for Old Dominion’s 2013 Norton Girault Poetry Prize
+
+## Other Writing
+
+### Documentation
+
+- [Implementing Docs as Code - Lessons and Benefits](https://jaxenter.com/docs-as-code-174941.html) - JAXenter - August 9, 2021
