--- netlify.toml
@@ -1,3 +1,0 @@
-[build]
-  publish = "dist"
-  command = "gridsome build"
