+++ nuxt.config.ts
@@ -1,0 +1,12 @@
+import { defineNuxtConfig } from 'nuxt/config'
+
+// https://v3.nuxtjs.org/api/configuration/nuxt.config
+export default defineNuxtConfig({
+  modules: [
+    '@nuxt/content',
+    '@nuxtjs/tailwindcss'
+  ],
+  css: [
+    '@fortawesome/fontawesome-svg-core/styles.css'
+  ]
+})
