+++ package.json
@@ -1,31 +1,25 @@
 {
-  "name": "andystevens.name",
   "private": true,
   "scripts": {
-    "develop": "gridsome develop",
-    "explore": "gridsome explore",
-    "build": "gridsome build"
-  },
-  "dependencies": {
-    "@gridsome/plugin-sitemap": "^0.4.0",
-    "@gridsome/remark-prismjs": "^0.5.0",
-    "@gridsome/source-filesystem": "^0.6.2",
-    "@gridsome/transformer-remark": "^0.6.4",
-    "gridsome": "^0.7.0",
-    "gridsome-plugin-gtag": "^0.1.10",
-    "gridsome-plugin-netlify-cms": "^1.0.9",
-    "gridsome-plugin-rss": "^1.2.0",
-    "netlify-cms": "^2.10.164",
-    "twojs-ts": "^0.7.0-13",
-    "vue2-hammer": "^2.1.2"
+    "build": "nuxt build",
+    "dev": "nuxt dev",
+    "generate": "nuxt generate",
+    "preview": "nuxt preview"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/vue-fontawesome": "^2.0.2",
-    "autoprefixer": "^10.3.4",
-    "postcss": "^8.3.6",
-    "tailwindcss": "^2.2.15"
+    "@nuxt/content": "^2.1.0",
+    "@nuxt/postcss8": "^1.1.3",
+    "autoprefixer": "^10.4.12",
+    "nuxt": "^3.0.0-rc.10",
+    "postcss": "^8.4.17",
+    "tailwindcss": "^3.1.8"
+  },
+  "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.2.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.0",
+    "@fortawesome/vue-fontawesome": "^3.0.1",
+    "@nuxtjs/tailwindcss": "^5.3.5",
+    "nuxt-hammer": "^2.1.3",
+    "two.js": "^0.8.10"
   }
 }
