+++ plugins/nuxt-hammer.js
@@ -1,0 +1,6 @@
+import { NuxtHammer } from 'nuxt-hammer'
+
+
+export default defineNuxtPlugin((nuxtApp) => {
+  nuxtApp.vueApp.use(NuxtHammer)
+})
