+++ readme.md
@@ -1,3 +1,19 @@
 # [andystevens.name](http://andystevens.name)
 
-My personal homepage.
+The personal homepage of Andy Stevens.
+
+## Dependenices
+
+- Tested with Node.js 18.0.0. This repo includes an [asdf](https://asdf-vm.com/) `.tool-versions` file and an [NVM](https://github.com/nvm-sh/nvm) `.nvmrc` file, so you if you have one of those tools installed you can run `asdf install` or `nvm use` in the project directory.
+- Nuxt.js 3
+- Tailwinds CSS
+
+## Installation
+
+```
+git clone git@github.com:andystevensname/andystevens.name.git
+cd andystevens.name
+npm install
+```
+
+To run locally, use `npm run dev`
