--- src/admin/config.yml
@@ -1,45 +1,0 @@
-backend:
-  name: github
-  repo: andystevensname/andystevens.name
-site_url: https://andystevens.name
-publish_mode: editorial_workflow
-media_folder: "static/uploads"
-public_folder: "/uploads"
-
-collections:
-  - name: "post"
-    label: "Posts"
-    folder: "blog"
-    create: true
-    slug: "{{year}}-{{month}}-{{day}}-{{fields.slug}}"
-    preview_path: "blog/{{fields.slug}}"
-    fields:
-      - label: "Title" 
-        name: "title"
-        widget: "string"
-      - label: "Slug"
-        name: "slug"
-        widget: "string"
-      - label: "Description"
-        name: "description"
-        widget: "string"
-      - label: "Publish Date"
-        name: "date"
-        widget: "datetime"
-        default: ""
-        date_format: "YYYY-MM-DD" # e.g. 24.12.2021
-        time_format: "HH:mm" # e.g. 21:07
-        format: "YYYY-MM-DD"
-        picker_utc: false
-      - label: "Publish Status"
-        name: "published"
-        widget: "select"
-        options:
-          - { label: "Not Published", value: "false" }
-          - { label: "Published", value: "true" }
-      - label: "Tags"
-        name: "tags"
-        widget: "list"
-      - label: "Body"
-        name: "body"
-        widget: "markdown"
