--- src/admin/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Netlify CMS</title>
-</head>
-<body>
-  <script src="index.js" type="module"></script>
-  <script>
-    CMS.registerPreviewStyle("https://andystevens.name/netlifycms.css");
-  </script>
-</body>
-</html>
