--- src/pages/404.vue
@@ -1,17 +1,0 @@
-<template>
-  <Layout>
-    <article class="px-4 relative">
-      <h1 class="w-full text-5xl mb-6 text-center">404</h1>
-      <h2>Page Not Found</h2>
-      <p>We couldn't find the page you were looking for. If you typed in the address, please double-check to make sure you typed it in correctly.</p>
-    </article>
-  </Layout>
-</template>
-
-<script>
-export default {
-  metaInfo: {
-    title: 'Page Not Found'
-  }
-}
-</script>
