--- src/pages/Writing.vue
@@ -1,43 +1,0 @@
-<template>
-  <Layout>
-    <article class="px-4 relative">
-      <h1 class="w-full text-5xl mb-6 text-center">Writing</h1>
-      <h2>Poetry</h2>
-      <h3>Publications</h3>
-        <ul>
-          <li>Eckerd Review - 2011</li>
-          <li><a href="http://www.magcloud.com/browse/issue/387253">Digital Americana</a> - Summer 2012 (published under Leroy van Stets)</li>
-          <li><a href="http://issuu.com/barelysouth/docs/barely_south_review_fall_2013">Barely South Review</a> - Fall 2013</li>
-          <li><a href="http://prickofthespindle.org/product/print-edition-issue-8/">Prick of the Spindle</a> - Issue 8, Spring/Summer 2015</li>
-          <li><a href="https://www.bridgeeight.com/poems-jack-freeman-andy-stevens/">Bridge Eight</a></li>
-        </ul>
-
-      <h3>Awards and Other Recognition</h3>
-        <ul>
-          <li>2009 ASPEC Mason Daly Poetry Award from Eckerd College</li>
-          <li>June 2013 Winner Tampa Bay Hotel / Plant Hall Beautiful-Weird-Cool-Spooky Old Place Writing Contest at University of Tampa</li>
-          <li>Finalist for Old Dominion’s 2013 Norton Girault Poetry Prize</li>
-        </ul>
- 
-      <h2>Other Writing</h2>
-      <h3>Documentation</h3>
-      <ul>
-          <li><a href="https://jaxenter.com/docs-as-code-174941.html">Implementing Docs as Code - Lessons and Benefits</a> - JAXenter - August 9, 2021</li>
-      </ul>
-    </article>
-  </Layout>
-</template>
-
-<script>
-export default {
-  metaInfo: {
-    title: 'Writing',
-    meta: [
-      {
-        name: 'description',
-        content: 'A full list of publications by Andy Stevens, including poetry and other writing.'
-      }
-    ]
-  }
-}
-</script>
