--- static/netlifycms.css
@@ -1,144 +1,0 @@
-html {
-    height: 100%;
-    min-height: 100%;
-    width: 100%;
-    position: relative;
-  }
-  
-body {
-    background-color: #212121;
-    height: 100%;
-    min-height: 100%;
-    max-width: 100%;
-    margin: 0;
-    position: relative;
-    padding: 0 1.5rem;
-    font-size: 100%;
-    font-size: 16px;
-    color: white;
-    font-family: Helvetica, Arial, sans-serif;
-}
-
-a, a:visited {
-    color: white;
-    text-decoration: none;
-    border-bottom: 2px dotted #00b2cc; 
-}
-
-a:hover, a:active {
-    color: #00b2cc;
-}
-
-ul {
-    list-style: square outside;
-}
-
-ol {
-    list-style: decimal outside;
-}
-
-h1 {
-    font-weight: bold;
-    font-size: 3rem;
-    width: 100%;
-    text-align: center;
-}
-
-h2 {
-    display: border-box;
-    width: 100%;
-    padding: 0;
-    margin: 0 0 1.5rem 0;
-    font-size: 1.875rem;
-    line-height: 3rem;
-    font-weight: bold;
-}
-
-h3 {
-    display: border-box;
-    width: 100%;
-    padding: 0;
-    margin: 0 0 1.5rem 0;
-    font-size: 1.25rem;
-    line-height: 1.5rem;
-    font-weight: bold;
-}
-
-p {
-    display: inline-block;
-    width: 100%;    
-    font-size: 1rem;
-    line-height: 1.5rem;
-    padding: 0;
-    margin: 0 0 1.5rem 0;
-}
-
-ul {
-    display: block;
-    width: 100%;
-    padding: 0 0 0 1rem;
-    margin-bottom: 1.5rem;
-}
-
-ol {
-    display: inline-block;
-    width: 100%;
-    position: relative;
-    padding-left: 1rem;
-    margin-bottom: 1.5rem;
-}
-
-ol li p, ul li p {
-    display: inline-block;
-    padding: 0;
-}
-
-ol li p:last-child, ul li p:last-child {
-    margin-bottom: 0;
-}
-
-blockquote {
-    max-width: 100%;
-    display: border-box;
-    border-left: 4px solid white;
-    padding: 0 1rem;
-    margin: 0 0 1.5rem 0;
-}
-
-blockquote p:last-child {
-    margin-bottom: 0;
-}
-
-blockquote ul:last-child, blockquote ol:last-child {
-  margin-bottom: 0;
-}
-
-blockquote > h2, h3, h4, h5, h6 {
-    border: 0;
-}
-
-code {
-  display: inline-block;
-  font-family: "Courier New", monospace;
-  background: black;
-  line-height: 1rem;
-  margin: 0 0 1.5rem 0;
-}
-
-p code {
-  margin-bottom: 0;
-}
-
-pre {
-  width: 100%;
-}
-
-.meta {
-    height: 1.5rem;
-    border-top: 1px dashed white;
-    font-size: .875rem;
-    line-height: 1.5rem;
-    display: border-box;
-    margin: 0 0 1.5rem 0;
-    width: 100%;
-}
