--- static/rss.xml
@@ -1,1 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><rss xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:content="http://purl.org/rss/1.0/modules/content/" xmlns:atom="http://www.w3.org/2005/Atom" version="2.0"><channel><title><![CDATA[Andy Stevens' Blog]]></title><description><![CDATA[Andy Stevens' Blog]]></description><link>https://andystevens.name</link><generator>RSS for Node</generator><lastBuildDate>Thu, 07 Oct 2021 22:36:48 GMT</lastBuildDate><atom:link href="https://andystevens.name/rss.xml" rel="self" type="application/rss+xml"/><item><title><![CDATA[On Forgetting]]></title><description><![CDATA[We spend so much time trying to remember, but what happens when we want to forget?]]></description><link>https://andystevens.name/blog/on-forgetting</link><guid isPermaLink="true">https://andystevens.name/blog/on-forgetting</guid></item></channel></rss>
